LayoutTests/imported/w3c/web-platform-tests/html/user-activation/propagation-sameorigin.html

<!DOCTYPE html>
<html>
  <head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <script src="resources/utils.js"></script>
  </head>
  <body></body>
  <script>
    async function createNestedIframes() {
      const iframe = await attachIframe("about:blank");
      const nested1 = await attachIframe(
        "about:blank#nested1",
        iframe.contentDocument
      );
      const nested2 = await attachIframe(
        "about:blank#nested2",
        nested1.contentDocument
      );
      return [iframe, nested1, nested2];
    }

    test(() => {
      assert_false(navigator.userActivation.isActive, "top isActive");
      assert_false(navigator.userActivation.hasBeenActive, "top hasBeenActive");
    }, "Initial user activation states");

    function checkIsActive(frame, expected) {
      const msg = `Frame ${frame.src} expects ${expected}`;
      const { isActive } = frame.contentWindow.navigator.userActivation;
      expected ? assert_true(isActive, msg) : assert_false(isActive, msg);
    }

    promise_test(async () => {
      const soloFrame = await attachIframe("about:blank#solo");
      const nestedFrames = await createNestedIframes();
      const iframes = [soloFrame, ...nestedFrames];

      // All frames userActivation's initially false
      for (const frame of iframes) {
        const { userActivation } = frame.contentWindow.navigator;
        assert_false(userActivation.isActive, `Frame ${frame.src}`);
        assert_false(userActivation.hasBeenActive, `Frame ${frame.src}`);
      }

      await test_driver.click(soloFrame);

      for (const frame of iframes) {
        checkIsActive(frame, true);
      }

      await consumeTransientActivation();

      assert_false(navigator.userActivation.isActive, "top isActive");
      assert_true(navigator.userActivation.hasBeenActive, "top hasBeenActive");
    }, "Tests that user activation propagates across same-origin frame boundary.");
  </script>
</html>
